updated db test to use hidden storage and db host info files
